processor_grpc/requirements.txt
grpcio==1.67.1
grpcio-tools==1.67.1
prometheus-client==0.21.1
